Розделил код категорий
