+++ controllers/category/index.js
@@ -1,0 +1,9 @@
+const getAll = require('./getAll');
+const add = require('./add');
+const remove = require('./remove');
+
+module.exports = {
+  getAll,
+  add,
+  remove,
+};
