+++ controllers/user/login.js
@@ -1,6 +1,7 @@
 const { BadRequest, Unauthorized } = require('http-errors');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
+require('dotenv').config();
 
 const { User } = require('../../models');
 const { joiLoginSchema } = require('../../models/user');
